# configs/swagger/v1.yaml
swagger: "2.0"
info:
  description: "悟空聊天IM API"
  version: "1.0.0"
  title: "悟空聊天IM API"
host: "49.235.59.182:8080"
...
schemes:
  - "http"
basePath: "/v1"
paths:
...
  /groups/{group_no}:
    get:
      tags:
        - "group"
      summary: "查询群信息"
      description: "查询群信息"
      operationId: "groupGet"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "group_no"
          type: string
          required: true
          description: "请求"
      responses:
        200:
          description: "成功"
          schema:
            $ref: '#/definitions/GroupResp'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: [] 
    put:
      tags:
        - "group"
      summary: "更新群组"
      description: "创建群组"
      operationId: "groupUpdate"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          type: string
          name: 'group_no'
          required: true
        - in: "body"
          name: "req"
          description: "如果只需要修改群名则只需要传name 同理修改公告只需要传notice"
          schema:
            $ref: '#/definitions/GroupUpdateReq'
      responses:
        200:
          description: "成功"
          schema:
            $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []  
...
  /groups/{group_no}/members:
    post:
      tags:
        - "group"
      summary: "添加群成员"
      description: "添加群成员"
      operationId: "memberAdd"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "group_no"
          type: "string"
          required: true
        - in: "body"
          name: "req"
          schema:
            $ref: '#/definitions/GroupmemberAddReq'
          description: "请求"
      responses:
        200:
          description: "成功"
          schema:
            $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
    delete:
      tags:
        - "group"
      summary: "移除群成员"
      description: "移除群成员"
      operationId: "memberRemove"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "group_no"
          type: "string"
          required: true
        - in: "body"
          name: "req"
          schema:
            $ref: '#/definitions/GroupmemberRemoveReq'
          description: "请求"
      responses:
        200:
          description: "成功"
          schema:
            $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []    
...
  /groups/{group_no}/forbidden/{on}:  
    post:
      tags:
        - "group"
      summary: "全员禁言"
      description: "全员禁言"
      operationId: "groupForbidden"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "group_no"
          type: "string"
          required: true
        - in: "path"
          name: "on"
          type: "number"
          description: "0.关闭禁言 1.开启禁言"
          required: true  
      responses:
        200:
          description: "成功"
          schema:
            $ref: '#/definitions/response'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
...
securityDefinitions:
  token:
    type: "apiKey"
    in: "header"
    name: "token"
    description: "用户token"
definitions:
...
  GroupResp:
    type: "object"
    description: "群信息"
    properties:
      group_no:
        type: string
        description: "群唯一编号"
      name: 
        type: string
        description: "群名称"  
      created_at:
        type: string
        description: "创建时间"
      version:
        type: integer
        description: "数据版本号"  
      mute:
        type: integer
        format: int
        description: "免打扰"
      top:  
        type: integer
        format: int
        description: "置顶"
      save:  
        type: integer
        format: int
        description: "保存到通讯录"  
      show_nick:  
        type: integer
        format: int
        description: "显示群成员昵称"    
      updated_at:
        type: string
        description: "更新时间"  
...
  GroupmemberRemoveReq:
    type: "object"
    properties:
      members:
        type: array
        items:
          type: string
        description: "成员uid集合"      
  GroupmemberAddReq:
    type: "object"
    properties:
      members:
        type: array
        items:
          type: string
        description: "成员uid集合"
...
  GroupUpdateReq:
    type: "object"
    properties:
      name:
        type: string
        description: "群名"
      notice:
        type: string
        description: "群公告"
      invite:
        type: number
        description: "是否开启确认邀请 0.否 1.是"  
...
  response:
    type: "object"
    properties:
      status:
        type: integer
        format: int
      msg:
        type: "string"